added footer and instructions for user

+++ src/app/tools-scupper-calculator/tools-scupper-calculator.component.ts
@@ -13,10 +13,21 @@ export class ToolsScupperCalculatorComponent implements OnInit {
 
   scupper: Scupper;
   scuppers: Scupper[];
+  choosenProject: string;
 
   instructionImg = 'assets/images/instructionscupper.png';
 
   constructor(private toolScupperService: ToolScupperService) { }
+
+  getProjectIfChoosen() {
+      this.choosenProject = localStorage.getItem('project');
+  }
+
+  checkIfProjectChoosen() {
+    if(this.choosenProject ===null) {
+      alert('if you want save scupper');
+    }
+  }
 
   checkScuppers(projectName: string,
     roofArea: string,
@@ -34,31 +45,36 @@ export class ToolsScupperCalculatorComponent implements OnInit {
   }
 
   findByProjectName(phrase: string): void {
-    // if (this.project) {
+    if (this.choosenProject) {
     this.toolScupperService.findByProjectName(phrase)
       .subscribe(scuppers => this.scuppers = scuppers);
-    // }
+    }
   }
 
   saveScupper(): void {
-    // if (this.project) {
+    if (this.choosenProject != null) {
     this.toolScupperService.saveScupper(this.scupper)
       .subscribe();
-    // }
+    } else {
+      alert('first select the project');
+    }
   }
 
   findAllScuppers(): void {
-    // if (this.project) {
+    if (this.choosenProject) {
     this.toolScupperService.findAll()
       .subscribe(scuppers => this.scuppers = scuppers);
-    // }
+    }
   }
 
   deleteScupper(scupper: Scupper): void {
-    // if (this.project) {
+    if (this.choosenProject) {
     this.toolScupperService.deleteScupperFromList(scupper)
-      .subscribe(scuppers => this.scuppers = scuppers);
-    // }
+      .subscribe(scuppers => {
+        this.scuppers = scuppers;
+        window.location.reload();
+      });
+    }
   }
 
   hideScuppers(): void {
@@ -66,11 +82,11 @@ export class ToolsScupperCalculatorComponent implements OnInit {
   }
 
   cleanAllScuppers(): void {
-    // if (this.project && confirm('Are you sure you want to clear all saved scuppers?')) {
+    if (this.choosenProject && confirm('Are you sure you want to clear all saved scuppers?')) {
       this.toolScupperService.clearAllScuppers()
         .subscribe(scuppers => this.scuppers = scuppers);
       this.findAllScuppers();
-    // }
+    }
   }
 
   scupperInitial(): Scupper {
@@ -92,6 +108,7 @@ export class ToolsScupperCalculatorComponent implements OnInit {
 
   ngOnInit(): void {
     this.scupper = this.scupperInitial();
+    this.getProjectIfChoosen();
   }
